Add Generator tests. Move all services test to designated place. Add basic support for ctest arguments pass for some tests.
