+++ computationalLib/test/GoogleCommon.h
@@ -8,6 +8,11 @@
 #include "common/MeshCreator.h"
 #include "common/commonTypes.h"
 
+
+#define SLEEP(dur) std::this_thread::sleep_for(dur)
+#define ASSERT_NEAR_REL(val1, val2, rel_error) ASSERT_NEAR(val1,val2,rel_error*val1/100)
+#define EXPECT_NEAR_REL(val1, val2, rel_error) EXPECT_NEAR(val1,val2,rel_error*val1/100)
+
 #define EXPECT_EXCEPTION_WITH_ARGS_BASE(statement, exception_type, expected_tuple, CMP) \
 do { \
     try { \
