+++ computationalLib/test/interegated/AMQPPublisherTests.cpp
@@ -1,0 +1,15 @@
+#include <gtest/gtest.h>
+
+
+#include "network_shared/AMQPPublisherService.h"
+
+TEST(ss,ss)
+{
+  ASSERT_TRUE(true);
+}
+
+
+int main(int argc, char **argv) {
+  ::testing::InitGoogleTest(&argc, argv);
+  return RUN_ALL_TESTS();
+}
