+++ utilLib/test/parallelUtils/ClockTests.cpp
@@ -1,14 +1,12 @@
 #include "parallelUtils/chronoClock.h"
+#include "../../computationalLib/test/GoogleCommon.h"
 #include <gtest/gtest.h>
 
 using chrono_clock::ChronoClockTemplate;
 using namespace timing;
 using namespace std::chrono_literals;
 
-#define SLEEP(dur) std::this_thread::sleep_for(dur)
-#define ASSERT_NEAR_REL(val1, val2, rel_error) ASSERT_NEAR(val1,val2,rel_error*val1/100)
-#define ASSERT_HISTORY_EQ(index, function) ASSERT_EQ(fff.call_history[index],(void *)function);
-#define COMMA ,
+
 
 using Ratio = std::milli;
 using DurationType = std::chrono::duration<int64_t, Ratio>;
